Refactor Middle High German and Middle English phonology modules
