+++ src/cltk/phonology/enm/stress.py
@@ -1,0 +1,210 @@
+"""
+
+"""
+
+import re
+from typing import List
+
+from cltk.stem.enm import stem
+
+from cltk.phonology.enm.syllabifier import SHORT_VOWELS, CONSONANTS, TRIPHTHONGS, DIPHTHONGS
+
+__author__ = ["Eleftheria Chatziargyriou <ele.hatzy@gmail.com>",
+              "Clément Besnier <clem@clementbesnier.fr>"]
+
+# Soundex replacement rules
+dict_SE = {
+    "p": "1",
+    "b": "1",
+    "f": "1",
+    "v": "1",
+    "gh": "1",
+    "sh": "2",
+    "t": "2",
+    "d": "2",
+    "s": "2",
+    "z": "2",
+    "r": "2",  # TODO
+    "k": "2",
+    "g": "2",
+    "w": "2",
+    "l": "3",
+    "m": "4",
+    "n": "4",
+    "r": "5",  # TODO
+}
+
+
+class MiddleEnglishStresser:
+    def __init__(self, syllabifier):
+        self.syllabifier = syllabifier
+
+    def stress(self, word, stress_rule="FSR") -> List:
+        """
+        :param word: word to stress
+        :param stress_rule: Stress Rule, valid options:
+
+            'FSR': French Stress Rule, stress falls on the ultima, unless
+             it contains schwa (ends with e), in which case the penult is
+            stressed
+
+            'GSR': Germanic Stress Rule, stress falls on the first syllable
+            of the stemm. Note that the accuracy of the function directly
+            depends on that of the stemmer.
+
+            'LSR': Latin Stress Rule, stress falls on the penult if its
+            heavy, else, if it has more than two syllables on the
+            antepenult, else on the ultima.
+
+        :return: A list containing the separate syllable, where the stressed
+            syllable is prefixed by ' . Monosyllabic words are left unchanged,
+            since stress indicates relative emphasis.
+
+        Examples:
+            >>> Word('beren').stresser(stress_rule = "FSR")
+            ['ber', "'en"]
+
+            >>> Word('prendre').stresser(stress_rule = "FSR")
+            ["'pren", 'dre']
+
+            >>> Word('yisterday').stresser(stress_rule = "GSR")
+            ['yi', 'ster', "'day"]
+
+            >>> Word('day').stresser(stress_rule = "GSR")
+            ['day']
+
+            >>> Word('mervelus').stresser(stress_rule = "LSR")
+            ["'mer", 'vel', 'us']
+
+            >>> Word('verbum').stresser(stress_rule = "LSR")
+            ['ver', "'bum"]
+        """
+
+        # Syllabify word
+        syllabified = self.syllabifier.syllabify(word)
+
+        # Check whether word is monosyllabic
+        if len(syllabified) == 1:
+            return syllabified
+
+        if stress_rule == "FSR":
+            # Check whether ultima ends in e
+            if syllabified[-1][-1] == "e":
+                return (
+                    syllabified[:-2]
+                    + ["'{0}".format(syllabified[-2])]
+                    + syllabified[-1:]
+                )
+
+            else:
+                return syllabified[:-1] + ["'{0}".format(syllabified[-1])]
+
+        elif stress_rule == "GSR":
+            # The word striped of suffixes
+            st_word = stem(word, strip_suf=False)
+            affix = word[: len(word) - len(st_word)]
+
+            # Syllabify stripped word and affix
+
+            syl_word = self.syllabifier.syllabify(st_word)
+
+            # Add stress
+            syl_word = ["'{0}".format(syl_word[0])] + syl_word[1:]
+
+            if affix:
+                affix = self.syllabifier.syllabify(affix)
+                syl_word = affix + syl_word
+
+            return syl_word
+
+        elif stress_rule == "LSR":
+            # Check whether penult is heavy (contains more than one mora)
+            if sum(map(lambda x: x in SHORT_VOWELS, syllabified[-1])) > 1:
+                return (
+                    syllabified[:-2]
+                    + ["'{0}".format(syllabified[-2])]
+                    + syllabified[-1:]
+                )
+
+            elif len(syllabified) > 2:
+                return (
+                    syllabified[:-3]
+                    + ["'{0}".format(syllabified[-3])]
+                    + syllabified[-2:]
+                )
+
+            else:
+                return syllabified[:-1] + ["'{0}".format(syllabified[-1])]
+
+    def phonetic_indexing(self, p="SE") -> str:
+        """
+        :param p: Specifies the phonetic indexing method
+                SE: Soundex variant for MHG
+
+        :return: Encoded string corresponding to the word's phonetic
+            representation
+        """
+
+        if p == "SE":
+            return self._Soundex()
+
+    def _Soundex(self):
+        """
+        The Soundex phonetic indexing algorithm adapted to ME phonology.
+
+        Algorithm:
+
+        Let w the original word and W the resulting one
+
+        1) Capitalize the first letter of w and append it to W
+
+        2) Apply the following replacement rules
+
+            p, b, f, v, gh (non-nasal fricatives) -> 1
+
+            t, d, s, sh, z, r, k, g, w (non-nasal alveolars and velars) -> 2
+
+            l (alveolar lateral) -> 3
+
+            m, n (nasals) -> 4
+
+            r (alveolar approximant) -> 5
+
+        3) Concetate multiple occurrences of numbers into one
+
+        4) Remove non-numerical characters
+
+        Notes:
+            /h/ was thought to be either a voiceless or velar fricative
+        when occurring in the coda with its most used grapheme being <gh>.
+        Those phonemes either disappeared, resulting in the lengthening
+        of preceding vowel clusters, or were developed into /f/ as evident
+        by modern spelling (e.g. 'enough': /ɪˈnʌf/ and 'though': /ðəʊ/)
+
+        Examples:
+            >>> Word("midel").phonetic_indexing(p="SE")
+            'M230'
+
+            >>> Word("myddle").phonetic_indexing(p="SE")
+            'M230'
+
+            >>> Word("might").phonetic_indexing(p="SE")
+            'M120'
+
+            >>> Word("myghtely").phonetic_indexing(p="SE")
+            'M123'
+        """
+
+        word = self.word[1:]
+
+        for w, val in zip(dict_SE.keys(), dict_SE.values()):
+            word = word.replace(w, val)
+
+        # Remove multiple adjacent occurences of digit
+        word = re.sub(r"(\d)\1+", r"\1", word)
+
+        # Strip remaining letters
+        word = re.sub(r"[a-zðþƿ]+", "", word)
+
+        # Add trailing zeroes and return
+        return (self.word[0].upper() + word + "0" * 3)[:4]
